# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MediTrade API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "عرض كل الخدمات",
                "description": "يعرض قائمة بجميع الخدمات",
                "operationId": "getServices",
                "responses": {
                    "200": {
                        "description": "قائمة الخدمات",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Service"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "خطأ داخلي"
                    }
                }
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "إنشاء خدمة جديدة",
                "description": "ينشئ خدمة جديدة ببيانات كاملة وصور",
                "operationId": "storeService",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title_en",
                                    "title_ar",
                                    "short_description_en",
                                    "short_description_ar",
                                    "full_description_en",
                                    "full_description_ar",
                                    "main_image",
                                    "header_image",
                                    "supplies_image",
                                    "main_image_alt_en",
                                    "main_image_alt_ar",
                                    "header_image_alt_en",
                                    "header_image_alt_ar",
                                    "supplies_image_alt_en",
                                    "supplies_image_alt_ar",
                                    "supplies_text_en",
                                    "supplies_text_ar"
                                ],
                                "properties": {
                                    "title_en": {
                                        "type": "string"
                                    },
                                    "title_ar": {
                                        "type": "string"
                                    },
                                    "short_description_en": {
                                        "type": "string"
                                    },
                                    "short_description_ar": {
                                        "type": "string"
                                    },
                                    "full_description_en": {
                                        "type": "string"
                                    },
                                    "full_description_ar": {
                                        "type": "string"
                                    },
                                    "meta_title": {
                                        "type": "string"
                                    },
                                    "meta_description": {
                                        "type": "string"
                                    },
                                    "main_image": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "header_image": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "supplies_image": {
                                        "type": "file",
                                        "format": "binary"
                                    },
                                    "main_image_alt_en": {
                                        "type": "string"
                                    },
                                    "main_image_alt_ar": {
                                        "type": "string"
                                    },
                                    "header_image_alt_en": {
                                        "type": "string"
                                    },
                                    "header_image_alt_ar": {
                                        "type": "string"
                                    },
                                    "supplies_image_alt_en": {
                                        "type": "string"
                                    },
                                    "supplies_image_alt_ar": {
                                        "type": "string"
                                    },
                                    "supplies_text_en": {
                                        "type": "string"
                                    },
                                    "supplies_text_ar": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1
                                        ],
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم إنشاء الخدمة",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Service"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "أخطاء التحقق من الصحة"
                    }
                }
            }
        },
...
        "/api/services/toggle-active/{id}": {
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "تفعيل/إلغاء تفعيل خدمة",
                "description": "يبدل حالة الخدمة بين مفعّلة وغير مفعّلة",
                "operationId": "toggleServiceActive",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "رقم تعريف الخدمة",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "تم تغيير حالة الخدمة",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service activation toggled"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Service"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "لم يتم العثور على الخدمة"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Service": {
                "title": "Service",
                "required": [
                    "id",
                    "title_en",
                    "title_ar"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title_en": {
                        "type": "string",
                        "example": "Service Title EN"
                    },
                    "title_ar": {
                        "type": "string",
                        "example": "عنوان الخدمة"
                    },
                    "short_description_en": {
                        "type": "string"
                    },
                    "short_description_ar": {
                        "type": "string"
                    },
                    "full_description_en": {
                        "type": "string"
                    },
                    "full_description_ar": {
                        "type": "string"
                    },
                    "meta_title": {
                        "type": "string"
                    },
                    "meta_description": {
                        "type": "string"
                    },
                    "slug_en": {
                        "type": "string"
                    },
                    "slug_ar": {
                        "type": "string"
                    },
                    "main_image": {
                        "type": "string"
                    },
                    "header_image": {
                        "type": "string"
                    },
                    "supplies_image": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ],
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}